+++ src/H5VLARRAY.h
@@ -23,14 +23,14 @@ herr_t H5VLARRAYmake( hid_t loc_id,
 
 herr_t H5VLARRAYappend_records( hid_t dataset_id,
                                 hid_t type_id,
-                                int nobjects,
+                                size_t nobjects,
                                 hsize_t nrecords,
                                 const void *data );
 
 herr_t H5VLARRAYmodify_records( hid_t dataset_id,
                                 hid_t type_id,
                                 hsize_t nrow,
-                                int nobjects,
+                                size_t nobjects,
                                 const void *data );
 
 herr_t H5VLARRAYget_info( hid_t   dataset_id,
