+++ tables/hdf5extension.pyx
@@ -1972,7 +1972,7 @@ cdef class VLArray(Leaf):
     return self.dataset_id, SizeType(nrecords), (SizeType(chunksize),), atom
 
 
-  def _append(self, ndarray nparr, int nobjects):
+  def _append(self, ndarray nparr, size_t nobjects):
     cdef int ret
     cdef void *rbuf
 
@@ -1995,7 +1995,7 @@ cdef class VLArray(Leaf):
 
     self.nrecords = self.nrecords + 1
 
-  def _modify(self, hsize_t nrow, ndarray nparr, int nobjects):
+  def _modify(self, hsize_t nrow, ndarray nparr, size_t nobjects):
     cdef int ret
     cdef void *rbuf
